Add compatibility with glibc 2.42
New glibc 2.42 changed the _rtld_global structure layout, so add code
to check the layout.

Also, refactor the code used to get symbol addresses and the dlsym
testcase.
