+++ include/Makefile.am
@@ -28,7 +28,8 @@ noinst_HEADERS = \
 	ld_rtld.h \
 	insn_queue.h \
 	insn_queue_lib.h \
-	minielf.h
+	minielf.h \
+	symbol_loader.h
 
 # Workaround a bug in Autoconf 2.69
 if CPU_X86_64
